requirements.txt
numpy
pandas
pytorch_lightning
pyyaml
torch
torchvision
wandb